Move fsStrategyLog to its own file and add a benchmark
Move fsEvaluationLogEntry and fsStrategyLog out of the already large
preemption.go into fair_sharing_log.go, as suggested in review.

Add BenchmarkRunFirstFsStrategy. It runs the first strategy with an
always-failing strategy so every candidate workload of one borrowing
ClusterQueue is evaluated, writes through zap's JSON encoder to a
counting sink, and reports the log volume emitted at V(4) against a V(2)
baseline. newFsLogFixture now takes testing.TB so the benchmark can
reuse it.
